Move examples and exercises of Future to exercise-scala-basic
These will be explained in the scala basic section.
