+++ exercise-scala-basic/src/main/scala/answer/Answer8.scala
@@ -1,4 +1,4 @@
-package exercise
+package answer
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent._
@@ -6,23 +6,18 @@ import scala.concurrent.duration._
 import scala.language.postfixOps
 import scala.util.{ Failure, Success }
 
-object Exercise4 extends App {
+object Answer8 extends App {
 
   def doubling(source: Future[Int]): Future[Int] = {
-    // (A) Future[Int] を受け取り、要素を2倍して返すメソッドを実装してみよう
-    ???
+    source.map(_ * 2)
   }
 
   def mapToDouble(source: Future[Int]): Future[Double] = {
-    // (B) Future[Int] を受け取り、浮動小数点数(Double)に変換して返すメソッドを実装してみよう
-    // 123.toDouble で浮動小数に変換できる
-    ???
+    source.map(_.toDouble)
   }
 
   def parseInt(source: Future[String]): Future[Int] = {
-    // (C) Future[String] を受け取り、整数にパースするメソッドを実装してみよう
-    // "123".toInt で整数に変換できる
-    ???
+    source.map(_.toInt)
   }
 
   // ここより下は簡単なテストコード
@@ -43,5 +38,4 @@ object Exercise4 extends App {
   Await.ready(parseFailure, 1 second)
 
   println("OK")
-
 }
