+++ exercise-scala-basic/src/main/scala/example/FutureExample.scala
