+++ exercise-scala-basic/src/main/scala/example/FutureFailureExample.scala
