+++ exercise-scala-basic/src/main/scala/example/FutureMapExample.scala
